Made `galactic_path_gen` more generic and renamed to `game_piece_path_gen`
