+++ zebROS_ws/src/behaviors/CMakeLists.txt
@@ -146,9 +146,9 @@ target_link_libraries(auto_node ${catkin_LIBRARIES})
 add_dependencies(auto_node ${behaviors_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 
 # 2021 actionlib servers
-add_executable(galactic_path_server src/galactic_path_gen.cpp)
-target_link_libraries(galactic_path_server ${catkin_LIBRARIES})
-add_dependencies(galactic_path_server ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+add_executable(game_piece_path_server src/game_piece_path_gen.cpp)
+target_link_libraries(game_piece_path_server ${catkin_LIBRARIES})
+add_dependencies(game_piece_path_server ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 
 # 2020 actionlib servers
 add_executable(num_powercells_pub_node src/num_powercells_pub_node.cpp)
@@ -325,7 +325,7 @@ install(TARGETS
 	climber_server
 	eject_server
 	elevator_server
-	galactic_path_server
+	game_piece_path_server
 	gotocolor_server
 	indexer_server
 	intake_cargo_server
