+++ zebROS_ws/src/behaviors/config/auto_mode_config.yaml
@@ -12,7 +12,7 @@ auto_mode_2: ["map_relative_barrel_racing_path"]
 auto_mode_3: ["map_relative_bounce_path"]
 auto_mode_4: ["map_relative_slalom_path"]
 auto_mode_5: ["dynamic_path"]
-auto_mode_6: ["galactic_search_path"]
+auto_mode_6: ["game_piece_pickup_path"]
 
 #auto_mode_1: ["start_intake", "path_to_trench_face_goal", "stop_intake", "shoot", "pause_2_sec"] #example
 #auto_mode_2: ["pause_2_sec","pause_5_sec"]
